requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
amqp==5.3.1
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.11
decorator==5.2.1
dill==0.3.8
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.18.0
flower==2.0.1
flwr==1.16.0
flwr-datasets==0.5.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
grpcio==1.71.0
huggingface-hub==0.29.3
humanize==4.12.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.34.0
iterators==0.0.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kombu==5.5.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
nest_asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.0
platformdirs==4.3.7
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==4.25.6
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.22.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.31.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
stack_data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
tomli==2.2.1
tomli_w==1.2.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
